backend/requirements.txt
flask==3.0.3
flask-cors==5.0.0
flask-marshmallow==1.2.1
flask-sqlalchemy==3.1.1
flask-caching==2.3.0
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
sqlalchemy==2.0.34
waitress==3.0.0
pytest==8.3.3
pytest-mock==3.14.0
coverage==7.6.1
setuptools==75.0.0
mock_alchemy==0.2.6
numpy==2.1.3
python-dateutil==2.9
isodate==0.7.2
celery==5.5.3
redis==6.4.0
tox==4.29.0
scipy==1.16.1
scikit-learn==1.7.2